Revert "{northbridge,soc,southbridge}: Don't use both of _ADR and _HID"
This reverts commit 01787608670adec26fcea48173e18395e51c790e.

AMD: Dropping the _HID of PCI root bus doesn't work well and people
started to notice the breakage.

Intel: These platforms have a devicetree switch to choose between PCI
and ACPI modes. In the former case we need _ADR, but in the latter _HID
as the PCI devices are hidden.

The conflicting use of _ADR and _HID still needs to be fixed before
we can bump our IASL version.

Change-Id: If7b52b9e8f2f53574849aa3fddfccfa016288179
Signed-off-by: Nico Huber <nico.huber@secunet.com>
Reviewed-on: https://review.coreboot.org/c/coreboot/+/37710
Reviewed-by: HAOUAS Elyes <ehaouas@noos.fr>
Reviewed-by: Kyösti Mälkki <kyosti.malkki@gmail.com>
Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>